Make user-activation and password-reset emails work again
I opted to simplify things, for now, and just hard-code the email templates into the app,
instead of exposing them through the Admin interface, as I was planning. That would be a
great feature, and it's one I still plan on tackling, but considering the time crunch,
it's not worth it at the moment. All the code is still there, for when we want to
implement that at a later time.

[completes #28584537] Set up Mailers for when a new User is created